+++ src/main/java/com/ens/repositories/DepartementRepository.java
@@ -10,7 +10,7 @@ import java.util.Optional;
 import com.ens.models.Departement;
 import com.ens.utils.DatabaseConnector;
 
-public class DepartementRepository implements BaseRepository<Departement, Integer> {
+public class DepartementRepository implements BaseRepository<Departement, Long> {
 
 	private static final String FIND_BY_ID_QUERY = "SELECT * FROM departements WHERE id = ?";
     private static final String FIND_BY_LONG_NAME_QUERY ="SELECT * FROM departements WHERE nom=? ";
@@ -20,10 +20,10 @@ public class DepartementRepository implements BaseRepository<Departement, Intege
     private static final String DELETE_QUERY = "DELETE FROM departements WHERE id = ?";
 
     @Override
-    public Optional<Departement> find(Integer id) throws SQLException {
+    public Optional<Departement> find(Long id) throws SQLException {
         try (Connection connection = DatabaseConnector.getConnection();
              PreparedStatement statement = connection.prepareStatement(FIND_BY_ID_QUERY)) {
-        	statement.setInt(1, id);
+        	statement.setLong(1, id);
             try (ResultSet result = statement.executeQuery()) {
                 if (result.next()) {
                     return Optional.of(cast(result));
@@ -34,10 +34,10 @@ public class DepartementRepository implements BaseRepository<Departement, Intege
         return Optional.empty();
     }
     
-    public Departement findById(Integer id) throws SQLException {
+    public Departement findById(Long id) throws SQLException {
         try (Connection connection = DatabaseConnector.getConnection();
              PreparedStatement statement = connection.prepareStatement(FIND_BY_ID_QUERY)) {
-        	statement.setInt(1,(int) id);
+        	statement.setLong(1,(Long) id);
             try (ResultSet result = statement.executeQuery()) {
                 if (result.next()) {
                 	 return cast(result);
@@ -102,7 +102,7 @@ public class DepartementRepository implements BaseRepository<Departement, Intege
     	boolean res;
         try (Connection connection = DatabaseConnector.getConnection();
              PreparedStatement statement = connection.prepareStatement(UPDATE_QUERY)) {
-        	statement.setInt(2, departement.getId());
+        	statement.setLong(2, departement.getId());
         	statement.setString(1, departement.getNom());
         	res= statement.executeUpdate() > 0;
             connection.close();
@@ -115,7 +115,7 @@ public class DepartementRepository implements BaseRepository<Departement, Intege
     	boolean res;
         try (Connection connection = DatabaseConnector.getConnection();
              PreparedStatement statement = connection.prepareStatement(DELETE_QUERY)) {
-        	statement.setInt(1, departement.getId());
+        	statement.setLong(1, departement.getId());
         	res= statement.executeUpdate() > 0;
             connection.close();
             return res;
@@ -124,7 +124,7 @@ public class DepartementRepository implements BaseRepository<Departement, Intege
 
     private Departement cast(ResultSet result) throws SQLException {
     	Departement departement = new Departement();
-    	departement.setId(result.getInt("id"));
+    	departement.setId(result.getLong("id"));
     	departement.setNom(result.getString("nom"));
         return departement;
     }
